Refactor `Channel`
See also: #53
